Now using a SPARQL endpoint for much more delightful speed
