+++ asclepius/asclepius/main.py
@@ -1,8 +1,10 @@
 import logging
-logging.basicConfig(level=logging.DEBUG)
 
 from flask import Flask, jsonify, request
 
+import dinto
+
+logging.basicConfig(level=logging.DEBUG)
 app = Flask(__name__)
 
 
@@ -29,20 +31,17 @@ def handle_invalid_usage(error):
     return response
 
 
-def to_labelled(query_result):
-    return [{'uri': uri, 'label': label} for (uri, label) in query_result]
-
-
 @app.route("/all_drugs", methods=['GET'])
 def drugs():
     """Return a list of all drugs listed in DINTO"""
-    return jsonify(to_labelled(dinto.all_drugs()))
+    resp = dinto.all_drugs()
+    return jsonify(dinto.all_drugs())
 
 
 @app.route("/all_ddis", methods=['GET'])
 def all_ddis():
     """Return a list of all drug-drug interactions listed identified in DINTO"""
-    return jsonify(to_labelled(dinto.all_ddis()))
+    return jsonify(dinto.all_ddis())
 
 
 @app.route('/ddis', methods=['POST'])
@@ -66,7 +65,7 @@ def ddis():
     except ValueError as e:
         raise InvalidUsage(str(e))
 
-    return jsonify(to_labelled(dinto_res))
+    return jsonify(dinto_res)
 
 
 @app.route("/ping")
